Fix incorrect handling of benchmark requests arising from short hard timeout when the experiment's search budget is small.
Hard timeout was not enough for EvoSuite runs when the search budget was less than 30s leading to uncertain and incorrect benchmark request cancellations. E.g., not all results of all runs would be recorded. To address this problem, we add a minimum hard timeout of 40 seconds and also add checks for thread interruptions in BufferedWriter configurations.